requirements.txt
Flask==3.0.2
SQLAlchemy==2.0.28
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
feedparser==6.0.11
beautifulsoup4==4.12.3
Flask-Login==0.6.3
bcrypt==4.1.2
openai>=1.14.2
arcadepy
# Database drivers
psycopg2-binary>=2.9.9  # PostgreSQL driver
# Okta integration
okta-jwt-verifier>=0.2.3
python-jose>=3.3.0
requests>=2.31.0
# Production web server
gunicorn>=21.2.0 